image-gen/requirements.txt
annotated-types==0.6.0
appdirs==1.4.4
apsw==3.43.1.0
async-timeout==4.0.3
attrs==23.2.0
autocommand==2.2.2
awscli==2.15.19
awscrt==0.19.19
beautifulsoup4==4.12.3
breezy==3.3.5
Brlapi==0.8.5
Brotli==1.1.0
brotlicffi==1.1.0.0
btrfsutil==6.7.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
configobj==5.0.8
coverage==7.4.1
crit==3.18
cryptography==42.0.5
css-parser==1.0.10
cssselect==1.2.0
discid==1.2.0
distlib==0.3.8
distro==1.9.0
dnspython==2.5.0
docutils==0.20.1
dulwich==0.21.7
fastbencode==0.2
fasteners==0.19
fastjsonschema==2.19.1
faust-cchardet==2.1.19
feedparser==6.0.11
filelock==3.13.1
gpg==1.23.2
html2text==2020.1.16
html5-parser==0.4.11
html5lib==1.1
idna==3.6
ifaddr==0.2.0
inflate64==1.0.0
inflect==7.0.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
jeepney==0.8.0
jmespath==1.0.1
lensfun==0.3.4
lit==17.0.6.dev0
lockfile==0.12.2
louis==3.29.0
lxml==5.1.0
Markdown==3.6
mechanize==0.4.8
merge3==0.0.14
meson==1.4.0
more-itertools==10.2.0
msgpack==1.0.5
multivolumefile==0.2.3
mutagen==1.47.0
netifaces==0.11.0
netsnmp-python==1.0a1
nftables==0.1
nicotine-plus==3.3.2
numpy==1.26.4
ordered-set==4.1.0
packaging==23.2
patiencediff==0.2.14
pdftotext==2.2.2
picard==2.11
pillow==10.2.0
platformdirs==4.2.0
ply==3.11
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pwquality==1.4.5
py7zr==0.21.0
pybcj==1.0.2
pycairo==1.26.0
pychm==0.8.6
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.12.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyGObject==3.46.0
PyJWT==2.8.0
pyppmd==1.1.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQt6==6.6.1
PyQt6-sip==13.6.0
PyQt6-WebEngine==6.6.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0
pyxdg==0.28
PyYAML==6.0.1
pyzstd==0.15.9
Reflector==2023.6.28.0.36.1
regex==2023.12.25
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scour==0.38.2
setproctitle==1.3.3
sgmllib3k==1.0.0
six==1.16.0
smbus==1.1
soupsieve==2.5
speedtest-cli==2.1.3
sshuttle==1.1.2
texttable==1.7.0
tomli==2.0.1
torbrowser-launcher==0.3.7
tqdm==4.66.2
trove-classifiers==2024.3.6
typing_extensions==4.10.0
unrardll==0.1.7
urllib3==1.26.18
urwid==2.6.8
validate==5.0.8
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
virtualenv==20.25.0
wcwidth==0.2.13
webencodings==0.5.1
xxhash==3.4.1
yt-dlp==2024.3.10
zeroconf==0.63.0
zstandard==0.22.0
